requirement.txt
Django==6.0.1
djangorestframework==3.16.1
inertia-django==1.2.0
django-vite==3.1.0
django-axes==8.1.0
django-unfold==0.44.0
django-admin-rangefilter==0.13.2
python-dotenv==1.2.1
colorlog==6.10.1
concurrent-log-handler==0.9.28
psutil==6.1.1

langchain==1.2.7
langchain-openai==1.1.7
langchain-core==1.2.7
langchain-community==0.4.1
langchain-chroma==1.0.0
langchain-classic==1.0.1
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0

chromadb==1.4.1
sentence-transformers==5.2.0
rank-bm25==0.2.2
pandas==3.0.0
openpyxl==3.1.5
pdfplumber==0.11.9
tabulate==0.9.0
requests==2.32.5

pdf2image==1.17.0
pytesseract==0.3.13
camelot-py==1.0.9
opencv-python-headless==4.11.0.86
marker-pdf[full]==1.10.2
